requirements.txt
python-telegram-bot>=20,<22
python-dotenv
aiosmtplib
